Add proper permissions in restore image script

+++ scripts/restore_img.sh
@@ -3,10 +3,11 @@
 LOCATION="/opt/lampp/htdocs/record-store/img"
 
 cd $LOCATION
-if [ $1 = '-d' ]
+if [ $# -gt 0 ] && [ $1 == "-d" ]
 then
     echo Clearing folder...
     sudo rm -r *
 fi
 sudo cp -r .backup/* .
+sudo chmod +777 * -R
 echo Images successfully restored!
